+++ honeycomb/__init__.py
@@ -1,14 +1,15 @@
-from .create_table import (check_schema_existence,
-                           check_table_existence,
-                           create_table_from_df)
+from .append_table import append_table
+from .create_table import create_table_from_df
 from .describe_table import describe_table
-from .run_query import run_query
-
+from .meta import get_table_storage_type_from_metadata
+from .querying import run_query
+from . import check
 
 __all__ = [
-    'check_schema_existence',
-    'check_table_existence',
+    'append_table',
+    'check',
     'create_table_from_df',
     'describe_table',
+    'get_table_storage_type_from_metadata',
     'run_query'
 ]