+++ honeycomb/config.py
@@ -1,9 +1,0 @@
-dtype_map = {
-    'object': 'STRING',
-    'int64': 'INT',
-    'float64': 'DOUBLE',
-    'bool': 'BOOLEAN',
-    'datetime64': 'DATETIME',
-    'timedelta': 'INTERVAL'
-    # 'category': None
-}