+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Table appending logic
+- Functionality to create a table with both table and column comments
+- Can create tables in Parquet format
+- Can generate filenames for new files being added to S3
+- Unit tests added for `create_table.py`, `append_table.py`, and `dtype_mapping.py`
+
+### Changed
+- Full table and column comments are required for creating a table outside of
+the experimental zone.
+- `create_table_from_df` and `append_table` now have overwrite protection.
+If a key is already present in S3, alarms will be raised unless explicitly
+told not to.
+
 ## [0.1.0] - 2020-05-11
 
 ### Added